Arreglos

+++ src/pages/UserHomeDetail/components/DataTable.jsx
@@ -56,6 +56,7 @@ const DataTable = ({ data }) => {
   };
 
   const notify = () => toast.success("Deleted successfully.");
+  const notifyUpdate = () => toast.success("Update successfully.");
 
   const filteredData = data?.filter((item) => {
     return Object.keys(filters).every((key) => {
@@ -76,17 +77,35 @@ const DataTable = ({ data }) => {
   });
 
   const handleDeleteRow = async (id) => {
+    Swal.fire({
+      title: `¿Deseas eliminar ${filters.receiptName}?`,
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Sí, eliminar recibo",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        try {
+          const data = axios.delete(`${apiUrl}/delete/${id}`);
+          getUserHouses();
+          notify();
+        } catch (error) {
+          console.log(error.message);
+        };
+      }
+    });
+  };
+
+  const handleEditRow = async (id) => {
+    setOpenModal(true);
     try {
-      const data = await axios.delete(`${apiUrl}/delete/${id}`);
+      const data = await axios.put(`${apiUrl}/update/${id}`);
       getUserHouses();
-      notify();
+      notifyUpdate();
     } catch (error) {
       console.log(error.message);
     }
-  };
-
-  const handleEditRow = (id) => {
-    setOpenModal(true);
   };
 
   return (
